pyproject.toml
[project]
name = "launch-inertia"
version = "0.1.0"
description = "Full-stack Django + Inertia.js + React application"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your@email.com" }
]

dependencies = [
    # Django Core
    "django>=6.0.2",
    "daphne>=4.1",
    "channels[daphne]>=4.0",

    # Inertia.js
    "inertia-django>=1.0",

    # Vite Integration
    "django-vite>=3.0",

    # Database
    "psycopg[binary,pool]>=3.2",
    "django-model-utils>=5.0",

    # CORS
    "django-cors-headers>=4.0",

    # Authentication
    "django-allauth>=65.0",
    "PyJWT>=2.8",
    "pyotp>=2.9",

    # Stripe
    "stripe>=8.0,<12.0",
    "dj-stripe>=2.10",

    # Background Tasks
    "celery[redis]>=5.4",
    "django-celery-beat @ git+https://github.com/celery/django-celery-beat.git@main",
    "django-celery-results>=2.5",
    "redis>=5.0",

    # Cache
    "django-redis>=5.4",
    "django-cacheops>=7.0",

    # Email
    "django-ses>=4.0",

    # Storage
    "django-storages[s3]>=1.14",
    "boto3>=1.35",
    "pillow>=11.0",

    # Admin
    "django-unfold>=0.40",
    "django-import-export>=4.0",

    # Utilities
    "python-dotenv>=1.0",
    "orjson>=3.10",
    "pydantic[email]>=2.9",
    "phonenumbers>=8.13",

    # Observability
    "django-structlog>=8.0",
    "structlog>=24.0",
    "sentry-sdk>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.9",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.24",
    "factory-boy>=3.3",
    "black>=24.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "django-stubs>=5.1",
    "ipython>=8.0",
    "django-debug-toolbar>=4.4",
    "django-extensions>=3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

# =============================================================================
# UV Configuration
# =============================================================================

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.9",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.24",
    "factory-boy>=3.3",
    "black>=24.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "django-stubs>=5.1",
    "ipython>=8.0",
    "django-debug-toolbar>=4.4",
    "django-extensions>=3.2",
]

[tool.uv]
# UV configuration options

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.ruff]
line-length = 88
target-version = "py312"
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]
strict = true
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings.development"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test_memory"
python_files = ["test_*.py", "*_test.py"]
python_paths = ["src"]
addopts = "-v --tb=short --reuse-db"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests",
]
